reorganized project directories and files
