+++ header/SIRCell.h
