+++ header/SIRModel.h
