+++ plotting/sir_line_plot.png
