+++ plotting/sir_stacked_area_plot.png
