+++ src/SIRModel.cpp
@@ -1,4 +1,4 @@
-#include "SIRModel.h"
+#include "../header/SIRModel.h"
 
 SIRModel::SIRModel(double b, double g, double timeStep, int steps)
     : beta(b), gammaRate(g), dt(timeStep), numSteps(steps) {}
